requirements.txt
python-dotenv~=0.21.1
git+https://github.com/Rapptz/discord.py#egg=discord.py
python-dateutil
typing-extensions
asyncpg
aiofiles